test: split oversized provider_readiness_part_002 to fix shard-8 line-limit guard
The Ollama URL repair commits grew test_provider_readiness_part_002.py to
1672 lines, tripping test_first_party_code_files_stay_under_line_limit
(MAX_FIRST_PARTY_FILE_LINES=1500) in coverage shard 8. Move the Ollama
URL-helper and overlay-profile credential tests into a new part_010,
leaving part_002 at 962 lines. Pure relocation; no test behavior or
coverage changes.
